Switch config_file to default False to fix PEP8 integration.
The default value for the config_file kwarg passed to
StyleGuide.__init__ is False; we were passing True which causes an
error when pep8 is installed, as PEP8 is expecting a filepath for the kwarg.

This isn't documented super well in PEP8; I think a better default would have
been None (or using an Optional or Maybe type, but that isn't pythonic).